pyproject.toml
[tool.poetry]
name = "llm-fastapi-service"
version = "0.1.0"
description = "Production-ready FastAPI service for Hugging Face LLM deployment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
PyYAML = "^6.0.1"
transformers = "^4.36.0"
torch = "^2.1.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
prometheus-client = "^0.19.0"
cachetools = "^5.3.2"
httpx = "^0.26.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
